update template and create view
rename file to keep pattern

update search filter parameters
