+++ landing/candidate/templates/candidate/candidate_page.html
@@ -58,4 +58,6 @@
 {% block extra_js %}
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.min.js" integrity="sha512-894YE6QWD5I59HgZOGReFYm4dnWc1Qt5NtvYSaNcOP+u1T9qYdvdihz0PPSiiqn/+/3e7Jo4EaG7TubfWGUrMQ==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
     <script src="{% static 'js/candidate/candidate.js' %}"></script>
+    <script src="{% static 'js/candidate/social.js' %}"></script>
+    <script src="{% static 'js/candidate/propositions.js' %}"></script>
 {% endblock %}
